New ItemIteratorRange class for iteration over buffers and subitems.
This makes it more natural to iterate over the items in a buffer or all
subitems of an item, because normal begin/end functions can be used
instead of templated ones. It also allows use of the range-for-loop.

+++ include/osmium/osm/area.hpp
@@ -144,8 +144,8 @@ namespace osmium {
          *
          * @returns Pair (number outer rings, number inner rings)
          */
-        std::pair<int, int> num_rings() const {
-            std::pair<int, int> counter { 0, 0 };
+        std::pair<size_t, size_t> num_rings() const {
+            std::pair<size_t, size_t> counter { 0, 0 };
 
             for (auto it = cbegin(); it != cend(); ++it) {
                 switch (it->type()) {
